initial android support
