+++ Cargo.toml
@@ -31,6 +31,9 @@ zengine_audio = { path = "crates/zengine_audio" }
 zengine_macro = { path = "crates/zengine_macro" }
 glam = "0.21.3"
 
+[target.'cfg(target_os = "android")'.dependencies]
+ndk-glue = { version = "0.7", features = ["logger"] }
+
 [dev-dependencies]
 bencher = "0.1.5"
