+++ crates/zengine_asset/src/lib.rs
@@ -2,6 +2,7 @@ mod asset_manager;
 mod assets;
 mod handle;
 mod io;
+mod io_task;
 
 use std::path::{Path, PathBuf};
 
@@ -32,11 +33,14 @@ impl AssetModule {
 impl Module for AssetModule {
     fn init(self, engine: &mut zengine_engine::Engine) {
         if let Some(asset_base_path) = self.asset_base_path {
-            #[cfg(not(target_arch = "wasm32"))]
+            #[cfg(all(not(target_arch = "wasm32"), not(target_os = "android")))]
             let asset_io = crate::io::FileAssetIo::new(asset_base_path);
 
             #[cfg(target_arch = "wasm32")]
             let asset_io = crate::io::WasmAssetIo::new(asset_base_path);
+
+            #[cfg(target_os = "android")]
+            let asset_io = crate::io::AndroidAssetIo::new(asset_base_path);
 
             engine.world.create_resource(AssetManager::new(asset_io));
         } else {
