+++ crates/zengine_audio/Cargo.toml
@@ -12,7 +12,7 @@ zengine_macro = { path = "../zengine_macro" }
 log = "0.4"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-rodio = { version = "0.15", default-features = false, features = ["vorbis", "flac", "wav"] }
+rodio = { version = "0.16", default-features = false, features = ["vorbis", "flac", "wav"] }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-rodio = { version = "0.15", default-features = false,  features = ["vorbis", "flac", "wav", "wasm-bindgen"] }
+rodio = { version = "0.16", default-features = false,  features = ["vorbis", "flac", "wav", "wasm-bindgen"] }
