+++ crates/zengine_input/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 serde = { version = "1.0", features = ["derive"] }
 fnv = "1.0.7"
 log = "0.4"
-winit = { version = "0.27.2", features = ["serde"] }
+winit = { version = "0.27.3", features = ["serde"] }
 gilrs = {version = "0.9.0", features = ["serde-serialize"]}
 zengine_ecs = { path = "../zengine_ecs" }
 zengine_engine = { path = "../zengine_engine" }
