+++ src/lib.rs
@@ -38,3 +38,6 @@ pub mod asset {
 
 extern crate zengine_macro;
 pub use zengine_macro::*;
+
+#[cfg(target_os = "android")]
+pub use ndk_glue;
